Update openshift.ks
Updated 'wget' URL (used by the node to pull the broker's public key) to include a parameter containing the node hostname.  The idea is to allow the broker to accumulate knowledge of nodes configured in this way.  That information may be used down the line for streamlined registration of nodes with the broker system.